High-Dimensional Mahalanobis balancing

High-Dimensional Mahalanobis balancing is a multivariate perspective of approximate covariate balancing method to estimate average treatment effect in causal inference. We use function hdMB to select variable then Mahalanobis balancing method will be applied to estimate ATE.

hdMB(
  x,
  treat,
  group1,
  group2 = NA,
  outcome,
  method = "MB",
  delta.space = c(0.1, 0.01, 0.001, 1e-04, 1e-05, 1e-06),
  GASMD.bound = 0.1,
  iterations = 1000
)

Arguments

x

covariates.

treat

treatment indicator vector.

group1

see Details.

group2

see Details, Default: NA.

outcome

outcome vector.

method

a string that takes values in "MB", "MB2", See Details, Default: 'MB'.

delta.space

a vector of candidate values for tuning parameter, See Details, Default: c(1e-1, 1e-2, 1e-3, 1e-4, 1e-5, 1e-6).

GASMD.bound

An upper bound for choosing the variables, See Details, Default: 0.1.

iterations

iteration times in optimization problem, Default: 1000.

Value

a hdMB object with the following attributes:

GMIM: Generalized Multivariate Imbalance Measure that defines in our paper.

GASMD: Generalized Absolute Standardized Mean Difference, See Details.

Details

group1,group2

To estimate \(E(Y (1))\) (average treatment effect for group 1), you need to set group1 = 1 and ignore group2. Similarly, To estimate \(E(Y (0))\) (average treatment effect for group 0), you need to set group1 = 0 and ignore group2.

To estimate average treatment effect on the control group \(E(Y (1) | T = 0)\), you need to set group1 = 1 and group2 = 0. Similarly, To estimate average treatment effect on the treated group \(E(Y (0) | T = 1)\), you need to set group1 = 0 and group2 = 1.

This function is feasible when there are more than two groups.

method can be a valid string, including

"MB": We choose the weighting matrix \({W}_1=[diag(\hat{\Sigma})]^{-1}\) where \(\hat{\Sigma}\) denotes sample covariance matrix.

"MB2": We choose the weighting matrix \({W}_2={[\hat{\Sigma}]^{-1}}\) where \(\hat{\Sigma}\) denotes sample covariance matrix.

delta.space grid of values for the tuning parameter, a vector of candidate values for the degree of approximate covariate balance. The default is c(1e-1, 1e-2, 1e-3, 1e-4, 1e-5, 1e-6).

GASMD.bound GASMD is an imbalance measure. We first standardize our covariate, then \(GASMD_{a,k} = (X_{1,a,k} + ... + X_{n_a,a,k}) / n_a - mean_k\) where \(X_{a}\) denotes the covariate in group a, k denotes the k-th dimension of covariate and \(n_a\) denotes the sample size in group a. Note that GASMD is different from ASMD. In our paper, we suggest using ASMD to select the variable. However, when there are more than two groups, ASMD is unsuitable. Therefore, we use GASMD instead of ASMD to select the variable.

Examples

#hdMB
##generating high-dimensional data
set.seed(0521)
data        <- si.data(sample.size = 200, dimension = 100)
dimension   <- dim(data$X)[2]

##estimating ATE##

##choosing variable
hdMB1       <- hdMB(x = data$X, treat = data$Tr, group1 = 1, outcome = data$Y, method = "MB")
hdMB1$GMIM
#>  [1] 1.009530e-09 4.698532e-10 3.585731e-10 2.898017e-10 2.963972e-10
#>  [6] 2.269770e-10 2.413735e-10 2.523061e-10 3.074013e-10 3.062586e-10
#> [11] 3.201180e-10 3.637801e-10 2.527609e-10 2.591048e-10 2.972363e-10
#> [16] 2.529676e-10 2.404731e-10 8.236534e-10 4.254837e-10 9.506346e-10
#> [21] 8.240780e-10 8.803717e-10 1.090486e-09 9.836493e-10 9.581269e-07
#> [26] 1.576464e-05
threshold   <- 24
index1      <- rank(hdMB1$GASMD) >= (dimension - threshold + 1)

##choosing variable
hdMB2       <- hdMB(x = data$X, treat = data$Tr, group1 = 0, outcome = data$Y, method = "MB")
hdMB2$GMIM
#>  [1] 1.013443e-09 4.165684e-10 3.060022e-10 2.989469e-10 2.019793e-10
#>  [6] 1.836889e-10 1.823804e-10 1.657426e-10 1.582662e-10 1.827693e-10
#> [11] 1.591133e-10 1.009574e-09 1.278928e-10 1.237531e-10 1.254830e-10
#> [16] 1.164060e-10 1.129730e-10 1.461194e-10 1.420197e-10 1.815353e-10
#> [21] 2.493475e-10 2.207225e-10 5.923506e-10 4.494290e-10 3.734073e-10
threshold   <- 25
index2      <- rank(hdMB2$GASMD) >= (dimension - threshold + 1)

##an estimate of ATE
result1     <- MB(x = data$X[,index1], treat = data$Tr, group1 = 1, outcome = data$Y)
result2     <- MB(x = data$X[,index2], treat = data$Tr, group1 = 0, outcome = data$Y)
result1$AT - result2$AT
#>          [,1]
#> [1,] 0.466609

##estimating ATC##
hdMB3       <- hdMB(x = data$X, treat = data$Tr, group1 = 1, group2 = 0, outcome = data$Y, method = "MB")
hdMB3$GMIM
#>  [1] 1.212616e-09 3.168743e-09 4.198361e-09 1.023991e-09 1.176201e-09
#>  [6] 4.389520e-09 4.859900e-09 3.890700e-04 6.189813e-03 6.180703e-03
#> [11] 7.681887e-03 7.083583e-03 1.239884e-02 1.377881e-02 1.402955e-02
#> [16] 1.302385e-02 1.343482e-02 2.059218e-02 2.117580e-02 2.557701e-02
#> [21] 2.467458e-02 2.760436e-02 2.801698e-02 2.695449e-02 2.584830e-02
#> [26] 2.602368e-02 2.610590e-02 2.627600e-02 2.534782e-02 2.792964e-02
#> [31] 2.809770e-02 3.014364e-02 3.033542e-02 3.028159e-02 2.969784e-02
#> [36] 3.086814e-02 3.078667e-02 2.996725e-02 2.919587e-02 2.908196e-02
#> [41] 2.835851e-02 2.828126e-02 2.793579e-02 2.727435e-02 2.668795e-02
threshold   <- 7
index3      <- rank(hdMB3$GASMD) >= (dimension - threshold + 1)
result3     <- MB(x = data$X[,index3], treat = data$Tr, group1 = 1, group2 = 0, outcome = data$Y)

##an estimate of ATC
result3$AT - mean(data$Y[data$Tr == 0])
#>          [,1]
#> [1,] 2.005467